Renamed even more changelogs to changelog (files too, python script...)
